Define mixin trait for bit pattern options
Properly define the constants sub-package

Scala has a slightly strange way of exporting packages that ONLY
define traits. Scala's `import` operation expects to receive an object
to import, but traits are not objects.
Thus, we need to define a package object that extends from the traits
that we want to expose.