Provide a legal control signal
This is the minimum required for compilation to happen. This is
because if the `ctrl_sigs` sequence is empty its type is Seq(Nothing),
which has no subtype. This causes a compile time error because the
value returned from the decode function invocation returns a
Seq(BitPat), where BitPat cannot be a subtype of Nothing.