Rework how individual sources are picked up by Make
Before we were using a wildcard expansion, which ended up including
the path preceding the file's name. Working with this additional piece
of information was annoying to work with. Instead, we use
Sub-Makefiles which add the file names in that directory to the
`TEST_SRCS` array. This provides a flat list of file names that need
to be compiled.

This may be complicated in the future if we have two source files that
have the exact same file name.