Remove the init.c baremetal utility library
The syscall.c library has both an _init and an _exit function, which
are both more robust than what the init.c library provides.

NHARTS can be removed because the init.c file I was using supposed
that there were multiple HARTs ona  single processor, which is not
true in this processor model. Perhaps in the future, where multi-hart
processors are being worked on my us, we will need to bring this back.