Connect the accelerator's interrupt line to legal instruction signal
This was initially quite confusing because the RoCCCoreIO interface
consists of two outputs and a single input. I initially thought the
exception line was INPUT for the accelerator. But I had the direction
wrong. Anything in the RoCCCoreIO class marked `Input` is input from
the main processor to the accelerator. The `Output` fields are for the
accelerator to talk back to the main core.