Properly decode the instruction to control signals
I was initially confused becasue rocket-chip uses several signal
registers, in addition to the decoding control signals. However, I
figured out/remembered this was because rocket-core is a pipelined
CPU, so it makes sense that each stage will require a set of control
signal registers, but the decode stage can treat them as just wires,
because they do not need to be held onto for too long.