Enable RoCC Interrupts
RoCC interrupts are not really documented anywhere beyond Rocket's MIP
CSR.
https://github.com/chipsalliance/rocket-chip/blob/b503f8ac28a497b2463ffbac84bfe66533ace0bb/src/main/scala/rocket/CSR.scala#L111

But, because of the way Chisel works, those CSR locations are fixed in
the order they are declared, with the top of the MIP Chisel class
being the highest-index bits of MIP. In addition, for simplicity, MIE
and MIP have the same interrupt conditions in the same indices. So...
the magic 0x2000 is actually going to set the RoCC MIE bit to 1,
enabling RoCC coprocessors to send interrupts back to the main core.