Add intermediate wires to hold data inputs to ALU
Wires in Chisel are slightly different than what we think about when
we normally think of wires. Wires can hold state! They may act like
buffers, or anything else we need to synthesize as.

This is different than a register (Reg), because registers have a way
to reset their state to some known value. Wires do not have this.

Use wires