Add IO bundle to DCacheFetcher, connected to RoCCCommandIO.mem bus
This separate module will be handling the interactions with the memory
system, so it is responsible for handling all the signals we may need
from the RoCCCommandIO class to request & receive data from the memory
subsystem.

Switch dmem_data to Wire

Wires can hold state over cycles too. "Wire"s can be synthesized as
registers. A "Reg" register requires a reset value and clock, which we
do not provide here.

Properly initialize dmem_data wire with a value