Use automatically expanding bit patterns for ALU functions
This way we only need to increment an integer to make the ALU
functional unit address change. It also prevents us from introducing
so many ALU functions that we need to increase the size of every
individual bit pattern. Instead, if we exceed the limit, we can just
increment the SZ_ALU_FN integer and add our new field.

It would be really nice to make an assertion or macro for this, so
that way you can never exceed the SZ_ALU_FN value anyways.