Big push of testing code for reading from memory & adding
The big difference between Wires and Registers is that Registers can
be updated (re-connected) within a when block for "conditional
assignment", which updates the register's stored contents. The Wire
does not allow this because it is a wire (duh...). It must always have
some kind of connection, and a conditional assignment breaks that
concept.