Simplify the data fetcher module
This module is intended to hold the FSM to fetch the necessary data
from memory. The FSM will fetch either one or two pieces of data and
return them, somehow.
Right now the results are returned on two separate data lines. But, it
may make more sense to initialize a "register file" Mem module at the
top-level and pass in the necessary signals to the relevant consumer
modules. In fact, a DecoupledIO may be better than just a simple
direct connection of Mem wires.