Allow transition to fetching data state on control path
This goes hand-in-hand with the previous commit. We need to make sure
we do not move to the next state unless the instruction is legal. In
the case that the funct7 opcode is 0x0, then the initial reset of the
processor causes the accelerator to move to the wrong state.