Set RoCC response based on register
Make sure blocking instructions properly block the main processor

We assume a default response case where the valid bit is false because
all instructions are blocking instructions right now. When the RoCC
processor is reset, this valid bit is reset to high, which causes the
accelerator to return as soon as it could.

This was not an issue before because the accelerator completed its
computation in just a single cycle, hiding this issue.

This ensures that we raise valid as long as we need to, returning it
back to false.B when we are done.