Store provided RoCC command in register until next instruction comes
Before, the completely combinational adder required zero clock cycles
to compute the result, meaning the command only needed to be stored
for one clock cycle, which was achieved by just assigning `cmd.bits`
to a wire. But now that the design takes more clock cycles, storing
the sent command for multiple cycles is required.

Only dequeue the RoCC instruction from RoCCIO queue when done

The deq at the "beginning" of the module prevents the current
instruction from being removed from the queue and the main processor
from sending us another one.

Technically, this could be removed, as we only get the bits for the
next instruction upon the current instruction being removed from the
queue with deq.