Replace use of RoCC command being valid with queue being fired
The `fire` method is slightly more general because the returned
boolean condition depends on both the ready and the valid bits of the
queue being true, rather than just the valid.
From the consumer's side (this accelerator), I don't think too much
should change in this case, but I could be wrong.