Use rocc command validity boolean to control exception generation
We can "control" this by using the rocc_cmd_valid register's value,
which ensures we do not accidentally send a response before we start
processing the sent instruction.

This change fixes an issue I introduced where the simulator will
properly error out on an illegal instruction, but the exception
handler gets "stuck" after the 14th register push (Register a4).
I have no idea why this is the case, but for some reason it fixes it.
This may be a quirk of the simulator or of my design, I am not sure
which.