Rework ALU class to be constructed as Module not CoreModule
We do not need to pass the implicit Parameters object either. This
makes sense, because the only thing the ALU can be parameterized by
right now is the length of the operands it operates over. We can
provide that from the module that instantiates the ALU, so the ALU
does not need this information either.

This also makes testing the design easier, as a config does not need
to be constructed that provides the implicit Parameters.