Fix simulator looping and hanging issues due to improper register use
The simulator would "die" when storing register a4 to the stack in the
interrupt handler! a4 "dying" was a combination of 2 things:
  1. A register is not needed to store the incoming RoCC instruction. Using a
wire (which allows for arbitrary delays/holding) is sufficient. I do
not know why a register vs. a wire was causing this behavior, but
making a direct connection avoids this issue entirely.
  2. Instructions which block must always return a result to the main
  processor, even if it will not be used (as is the case with an interrupt).

The simulator would loop in the exception handler when raising an
exception! This looping happened because the response I was sending
was always being constructed and sent. The response data was connected
to the ALU's output, which was outputting 0 at the time, and the
register the response should be stored in was just used directly from
the given RoCC command. This caused the main processor to pull that
data and store it in a15, which corresponded to `int c` in the test
program. When using that value, weird stuff would happen and the real
data would be overwritten by that 0 from the RoCC ALU.

The looping of the simulator was fall-out of the first part, where a
result ALWAYS needs to be sent back to the main processor when the
RoCC instruction sent is one that requires a result, even when the
result is not used, as is the case when an exception is raised.