Add companion object to CtrlSigs to convert bit patterns to signals
This is particularly useful for unit tests, as we want to convert a
sequence/list of bit patterns to other types, WITHOUT instantiating
hardware, like Wire or IO, as hardware cannot be created inside a
test.
The only hardware that can be created for a test is the design under
test.