Remove explicit Output direction on L1 D$ fetcher
This fixes an error from Chisel about the top-level VCodeAccel module
not being able to write to the DCache fetcher's decoupled request's
ready bit. Exact error message included:
  chisel3.internal.ChiselException: Connection between
  sink (DCacheFetcher.io.req.ready: IO[Bool]) and
  source (VCodeAccel.io.mem.req.ready: IO[Bool]) failed @: io.req.ready
  in DCacheFetcher cannot be written from module VCodeAccel.

This happens because putting Output on an IO component makes ALL
components in the bundle be outputs, which does not work for certain
things.