Pass control signals to control when to set L1 D$ valid bit
If we do not pass something that checks for the legality of the RoCC
instruction, the coprocessor will essentially generate requests to the
cache for memory that may be invalid. As discussed on Google
Groups[1], RoCC L1 D$ requests are not allowed to cause exceptions.

An exception in this case is a valid request to an inaccessible
address. Because the error was occurring at the VERY beginning of the
simulation, it is likely the accelerator was making a requests to
address 0, which is not allowed immediately after starting the
processor.

[1] https://groups.google.com/u/1/g/chipyard/c/OIbByFj6cZ0?hl=en