Simplify control unit's presentation of internal state to outside
Here we use a bunch of boolean comparisons that are always hooked up
and compare against the current state of the control unit's FSM,
rather than manual assignment to the various output booleans.

This reduces the amount of mental strain on the designer, and may
improve the timing of the control unit. These booleans update at the
exact same time the FSM changes state, rather than having to wait
another clock cycle for the internal boolean registers to switch to
their new values.