Store register to send response to in a register
This breaks things that must complete in a single cycle, because
Regs are only updated on rising edges. I would prefer a transparent
latch, so that the changes to the register are immediate.

But, this fixes multi-cycle coprocessor computations, with the result
returned to the correct register at the correct time.