Remove use of Queue and use io's DecoupledIO directly
The way we were using the Queue constructor turns a ReadyValidIO
bundle into a DecoupledIO one. But, the RoCC IO is already a
DecoupledIO bundle, so there is no need to do this.

The queue also added a 1 cycle latency, because of the hand-shaking
between the valid/ready bits from the incoming IO and the Queue. This
made this drastically more confusing and difficult, as things did not
behave the way we think they should

This ALSO fixes the issue where the Verilator simulator would just
hang saving register a4 during the interrupt handler, exactly 100
cycles after the RoCC command was issued. It turns out, when the
interrupt fires, the response register MUST be the exact same as what
was sent with the command.

This now introduces the possibility to "retire" the command by setting
the validity flag to be false, which could be a very powerful idea.