Define MStatus in terms of saved RoCC command
This is instead of using ANOTHER register (which would introduce
another cycle of latency). Because we are saving the command, we are
assured that the data we want will be alive for as long as we need
it.

This assumes someone does not "retire" the sent command earlier than
intended.