Rework converting lists of bit patterns to control signals
This is never touched by designs that are elaborated, but only by
unit tests.

This change is required for functional units within the ALU that have
unknown addresses (FN_X, for instance) which is defined as
BitPat.dontCare(). Unknown bit patterns have no way to convert to
anything that is a subtype of Data (Bool, UInt, SInt, etc.) because
there is no way to convert unknowns to known-value bits.