Fetcher must wait for response based on calculated tag value
The version of the fetcher before this change would error when
fetching a single value from a memory location that generated a tag
value not at index 0.

This is because the single value from that location would have
reqs_sent=0, causing us to loop around, submit another memory request
with the same tag number, but never mark the original reqs_sent=0
index as completed.