Add new state to control unit for writing back to memory
This new state is inserted between the execute and response states. We
can only write back to memory after executing, and once the
accelerator is done, it must send a response that notifies the main
core of the accelerator's completion (if the RoCC instruction required
that a response be sent back).