Allow writes to unmark their wait-for-response flag
Loads and stores behave the same way for io.resp.valid, meaning the
valid response flag is raised when the write is completed, usually by
~2 cycles later in the simplest case). However, io.resp.bits.data gets
set to 0 when the write completes. The has_data boolean flag is only
raised for reads/loads, as a write/store does NOT produce data.

By removing the response handler's dependence on has_data, we can now
respond to any completed tags, regardless if the cache subsystem
believes new data has been produced or not.