Allow accelerator to return a status flag after instruction completes
You MUST send a response for an instruction that blocks the main core.
But, since we are no longer returning the result of our single-element
integer vector addition directly to the CPU, we can use this xLen
block of data for anything.

It could be useful to have this track as a status code, for something
that is not exception-raising, but could communicate information back.