Use counter for tag rather than address bits
This is important for the ordering of elements within the fetcher's
internal Mem structure. These need to be in the same order in the data
fetcher's memory as in actual memory, because otherwise different runs
of the data fetcher with different base addresses will confuse place
elements of each vector in a different order, which will cause
problems for pair-wise operations.

NOTE: The io.start on should_send_request is required! If it were
missing, the should_send_request, and the io.req.valid flag would
always be true when executing. When simulating, this leads to a
violation of the memory subsystem's internal state, which raises a
$fatal exception in the Verilog generated from rocket's
SimpleHellaCacheIF.scala file.

Debugging prints when writing use tag value