Use intermediate wire to name result of subtraction of operand counter
We use a "bottoming" unsigned integer subtraction. This means if n > x
and we perform x-n, instead of getting a negative number, we get 0.

When we perform the subtraction, if the result were to go
"negative" (in reality this is a VERY positive number because the
integer is treated as unsigned), we set to 0, as this means we reached
the end of the vector and need to signal as such.