Fix ALU data being written 1 cycle too early
By connecting the data_fetcher's input to alu_out, which is a
register, we latch/capture the ALU's output in the top-level. This
allows us to sit around holding the results for as many cycles as we
need until the control unit says writing should begin.

This wire redirection looks a little silly, but it makes sense for the
upcoming vector addition reduction.