ALU's output no longer has valid flag, making it "stateless"
The control unit is now the one who determines if the operation is
completed. Right now, all the operations we care about seem to be able
to be implemented in a combinational way. I am not sure if this is
really the case for the future, but we will jump that hurdle once we
get there.

The "stateless" ALU makes things slightly simpler, because the ALU now
no longer signals back to the control unit. The control unit is the
only one in charge of moving into and out of the execution state.

NOTE: Currently, this only works with purely combinatorial operations.