Add requirement that the batchSize be a power of 2
This check is done by Scala during the early stages of elaboration,
which makes these violations get caught quick.

By the point this is checked, all parameters for all modules are known
to the Chisel compiler. These parameters come from the individual
instantiating an instance of this module. Violating this requirement
means the individual who created an instance of the VCode accelerator
provided an incorrect value by hand somewhere.