Add flake.nix that mirrors shell.nix
The biggest benefit we get here is that we can have ALL machines track
the same version of nixpkgs, without needing to resort to us writing
the Nix to manually fetch the nixpkgs tarball of the commit we want.

NOTE: You likely want this to track the same version of nixpkgs as is
used in the flake.nix for Chipyard, but there is no strict requirement
for that!