Hack in User-mode delegation
If the exception that is raised is one we care about (illegal
instruction in this case) AND we have already enabled M->S delegation,
then we further delegate the exception to U-mode.

NOTE: This is a half-baked way to do this! The processor still
"believes" this was a delegation to S-mode, so you MUST ensure you set
up the relevant S-mode registers too (stvec for example)!