User delegation test uses exit(0) to signal successful simulation
handle_trap reads from an xCAUSE register, regardless of the actual
state of the processor, which means that if the user-delegation
exception handler actually handles the exception, handle_trap would
still think something is wrong and report that.

Since this is a small test case, we should be fine just exiting at
this point. We will probably catch other problems using different
means.