Enable assertion in DCacheFetcher about state of inputs
A simple comparison of amountData is invalid because amountData WILL
go to 0 during execution, namely when we have finally exhausted the
vector and are done. We only care that amountData > 0 when we are
going to START executing the cache requests, namely when io.start is
true. Conjunction is too restrictive, but implication is exactly what
we want.