Rework MAX/MIN_REDUCE to use .reduce & built-in .max/.min methods
Chisel defines max and min functions for their numerical data types,
which we can leverage instead of using just the comparison operators.
This produces the same result in Verilator simulation, but it offers
the compilers more opportunities to generate & optimize the hardware
because of the additional semantic information we provide.

This also has a side-effect of looking nearly identical to the
reduction function we use for all the other reduction operators too,
which is a nice benefit for legibility.