Reduce shift slicing because VCODE spec states shifts have max shift range
The VCODE spec says that the values used as the shift amount in both
left and right shifts cannot be too large. In our implementation where
everything is 64 bits wide, we rely on shamt<64. This means we only
need 6 bits to represent the shift amount.

This is particularly important during elaboration because Chisel &
CIRCT will generate a SystemVerilog implementation that _could_
accomodate a 19-bit shift, which meant that we ended up generating
output buses that were 500k bits wide, which hurt things during FPGA
synthesis.