Replace all premature uses of << for *
Many times when we want to do multiplication, I (Karl) prematurely
optimized and used left-shifts instead of multiplications. But looking
at the generated code for compile-time multiplications, such a thing
already happens for us.

So, replace the premature << with the semantically clearer * operator,
and rely on Chisel to detect the fact we are doing multiplication by a
power of 2 and have it generate the appropriate left shifts.